+++ src/coupler/ocean/PicoGeometry.cc
@@ -24,6 +24,8 @@
 #include "pism/util/pism_utilities.hh"
 #include "pism/util/petscwrappers/Vec.hh"
 
+#include "pism/coupler/util/options.hh"
+
 namespace pism {
 namespace ocean {
 
@@ -32,6 +34,7 @@ PicoGeometry::PicoGeometry(IceGrid::ConstPtr grid)
       m_continental_shelf(grid, "pico_contshelf_mask", WITHOUT_GHOSTS),
       m_boxes(grid, "pico_box_mask", WITHOUT_GHOSTS),
       m_ice_shelves(grid, "pico_shelf_mask", WITHOUT_GHOSTS),
+      m_basin_mask(m_grid, "basins", WITH_GHOSTS),
       m_distance_gl(grid, "pico_distance_gl", WITH_GHOSTS),
       m_distance_cf(grid, "pico_distance_cf", WITH_GHOSTS),
       m_ocean_mask(grid, "pico_ocean_mask", WITH_GHOSTS),
@@ -46,11 +49,11 @@ PicoGeometry::PicoGeometry(IceGrid::ConstPtr grid)
   m_ice_rises.metadata().set_string("flag_meanings",
                                      "ocean ice_rise continental_ice_sheet, floating_ice");
 
-  m_tmp_p0 = m_tmp.allocate_proc0_copy();
-}
+  m_basin_mask.set_attrs("climate_forcing", "mask determines basins for PICO",
+                         "", "", "", 0);
+  m_n_basins = 0;
 
-PicoGeometry::~PicoGeometry() {
-  // empty
+  m_tmp_p0 = m_tmp.allocate_proc0_copy();
 }
 
 const IceModelVec2Int &PicoGeometry::continental_shelf_mask() const {
@@ -69,18 +72,32 @@ const IceModelVec2Int &PicoGeometry::ice_rise_mask() const {
   return m_ice_rises;
 }
 
+const IceModelVec2Int &PicoGeometry::basin_mask() const {
+  return m_basin_mask;
+}
+
+void PicoGeometry::init(const IceModelVec2CellType &cell_type) {
+
+  ForcingOptions opt(*m_grid->ctx(), "ocean.pico");
+
+  m_basin_mask.regrid(opt.filename, CRITICAL);
+
+  m_n_basins = static_cast<int>(m_basin_mask.max()) + 1;
+
+  m_n_basin_neighbors.resize(2*m_n_basins);
+  get_basin_neighbors(cell_type, m_basin_mask, m_n_basin_neighbors);
+}
+
 /*!
  * Compute masks needed by the PICO physics code.
  *
  * After this call box_mask(), ice_shelf_mask(), and continental_shelf_mask() will be up
  * to date.
  */
-void PicoGeometry::update(const IceModelVec2S &bed_elevation, const IceModelVec2CellType &cell_type) {
+void PicoGeometry::update(const IceModelVec2S &bed_elevation,
+                          const IceModelVec2CellType &cell_type) {
+
   bool exclude_ice_rises = m_config->get_flag("ocean.pico.exclude_ice_rises");
-
-  int n_boxes = m_config->get_number("ocean.pico.number_of_boxes");
-
-  double continental_shelf_depth = m_config->get_number("ocean.pico.continental_shelf_depth");
 
   // these three could be done at the same time
   {
@@ -89,6 +106,7 @@ void PicoGeometry::update(const IceModelVec2S &bed_elevation, const IceModelVec2
     compute_ocean_mask(cell_type, m_ocean_mask);
 
     compute_lakes(cell_type, m_lake_mask);
+
   }
 
   // these two could be optimized by trying to reduce the number of times we update ghosts
@@ -101,12 +119,27 @@ void PicoGeometry::update(const IceModelVec2S &bed_elevation, const IceModelVec2
     compute_distances_cf(m_ocean_mask, m_ice_rises, exclude_ice_rises, m_distance_cf);
   }
 
-  // these two could be done at the same time
+  // computing ice_shelf_mask and box_mask could be done at the same time
   {
     compute_ice_shelf_mask(m_ice_rises, m_lake_mask, m_ice_shelves);
+    auto n_shelves = static_cast<int>(m_ice_shelves.max()) + 1;
 
-    compute_continental_shelf_mask(bed_elevation, m_ice_rises, continental_shelf_depth, m_continental_shelf);
+    std::vector<int> cfs_in_basins_per_shelf(n_shelves*m_n_basins, 0);
+    std::vector<int> most_shelf_cells_in_basin(n_shelves, 0);
+    identify_calving_front_connection(cell_type, m_basin_mask, m_ice_shelves, n_shelves,
+                                      most_shelf_cells_in_basin, cfs_in_basins_per_shelf);
+
+    split_ice_shelves(cell_type, m_basin_mask, m_n_basin_neighbors,
+                      most_shelf_cells_in_basin, cfs_in_basins_per_shelf, n_shelves,
+                      m_ice_shelves);
+
+    double continental_shelf_depth = m_config->get_number("ocean.pico.continental_shelf_depth");
+
+    compute_continental_shelf_mask(bed_elevation, m_ice_rises, continental_shelf_depth,
+                                   m_continental_shelf);
   }
+
+  int n_boxes = static_cast<int>(m_config->get_number("ocean.pico.number_of_boxes"));
 
   compute_box_mask(m_distance_gl, m_distance_cf, m_ice_shelves, n_boxes, m_boxes);
 }
@@ -129,7 +162,7 @@ static void relabel(RelabelingType type,
 
   IceGrid::ConstPtr grid = mask.grid();
 
-  int max_index = mask.range().max;
+  int max_index = static_cast<int>(mask.range().max);
 
   if (max_index < 1) {
     // No components labeled. Fill the mask with zeros and quit.
@@ -146,7 +179,7 @@ static void relabel(RelabelingType type,
       for (Points p(*grid); p; p.next()) {
         const int i = p.i(), j = p.j();
 
-        int index = mask(i, j);
+        int index = mask.as_int(i, j);
 
         if (index > max_index or index < 0) {
           throw RuntimeError::formatted(PISM_ERROR_LOCATION, "invalid component index: %d", index);
@@ -176,7 +209,7 @@ static void relabel(RelabelingType type,
     int biggest_component = 0;
     for (unsigned int k = 0; k < area.size(); ++k) {
       if (area[k] > area[biggest_component]) {
-        biggest_component = k;
+        biggest_component = static_cast<int>(k);
       }
     }
 
@@ -221,7 +254,11 @@ void PicoGeometry::label_tmp() {
   try {
     if (m_grid->rank() == 0) {
       petsc::VecArray mask_p0(*m_tmp_p0);
-      label_connected_components(mask_p0.get(), m_grid->My(), m_grid->Mx(), false, 0.0);
+      label_connected_components(mask_p0.get(),
+                                 static_cast<int>(m_grid->My()),
+                                 static_cast<int>(m_grid->Mx()),
+                                 false,
+                                 0.0);
     }
   } catch (...) {
     rank0.failed();
@@ -229,10 +266,6 @@ void PicoGeometry::label_tmp() {
   rank0.check();
 
   m_tmp.get_from_proc0(*m_tmp_p0);
-}
-
-static bool edge_p(int i, int j, int Mx, int My) {
-  return (i == 0) or (i == Mx - 1) or (j == 0) or (j == My - 1);
 }
 
 /*!
@@ -260,7 +293,7 @@ void PicoGeometry::compute_lakes(const IceModelVec2CellType &cell_type, IceModel
     if (cell_type.ocean(i, j)) {
       m_tmp(i, j) = 1.0;
 
-      if (edge_p(i, j, Mx, My)) {
+      if (grid_edge(*m_grid, i, j)) {
         m_tmp(i, j) = 2.0;
       }
     } else {
@@ -345,7 +378,8 @@ void PicoGeometry::compute_ice_rises(const IceModelVec2CellType &cell_type, bool
  * 2 - ice-free areas with bed elevation > threshold, connected to the continental ice sheet
  */
 void PicoGeometry::compute_continental_shelf_mask(const IceModelVec2S &bed_elevation,
-                                                  const IceModelVec2Int &ice_rise_mask, double bed_elevation_threshold,
+                                                  const IceModelVec2Int &ice_rise_mask,
+                                                  double bed_elevation_threshold,
                                                   IceModelVec2Int &result) {
   IceModelVec::AccessList list{ &bed_elevation, &ice_rise_mask, &m_tmp };
 
@@ -472,6 +506,200 @@ void PicoGeometry::compute_ocean_mask(const IceModelVec2CellType &cell_type, Ice
   result.copy_from(m_tmp);
 }
 
+/*!
+ * Find the two neighboring basins by checking for the basin boundaries on the ice free ocean.
+ * Could there be more than two neighbors? Should we better identify the intersection at the coastline?
+ */
+void PicoGeometry::get_basin_neighbors(const IceModelVec2CellType &cell_type,
+                                       const IceModelVec2Int &basin_mask,
+                                       std::vector<int> &result) {
+
+  // additional vectors to allreduce efficiently with IntelMPI
+  std::vector<int> result1(2 * m_n_basins, 0);
+
+  IceModelVec::AccessList list{ &cell_type, &basin_mask };
+
+  for (Points p(*m_grid); p; p.next()) {
+    const int i = p.i(), j = p.j();
+
+    int b = basin_mask.as_int(i, j);
+    if (result[2 * b] == 0 or result[2 * b + 1] == 0) {
+
+      auto B = basin_mask.int_star(i, j);
+      auto M = cell_type.int_star(i, j);
+      int bn = 0; // neighbor basin id
+
+      if (M.ij == MASK_ICE_FREE_OCEAN and
+          ((M.n == MASK_ICE_FREE_OCEAN and B.n != b) or
+           (M.s == MASK_ICE_FREE_OCEAN and B.s != b) or
+           (M.e == MASK_ICE_FREE_OCEAN and B.e != b) or
+           (M.w == MASK_ICE_FREE_OCEAN and B.w != b))) {
+
+        if (M.n == MASK_ICE_FREE_OCEAN and B.n != b) {
+          bn = B.n;
+        } else if (M.s == MASK_ICE_FREE_OCEAN and B.s != b) {
+          bn = B.s;
+        } else if (M.e == MASK_ICE_FREE_OCEAN and B.e != b) {
+          bn = B.e;
+        } else if (M.w == MASK_ICE_FREE_OCEAN and B.w != b) {
+          bn = B.w;
+        }
+
+        if (result[2 * b] == 0) {
+          result[2 * b] = bn;
+        } else if (result[2 * b + 1] == 0 and result[2 * b] != bn) {
+          result[2 * b + 1] = bn;
+        }
+      }
+    }
+  }
+
+  GlobalSum(m_grid->com, result.data(), result1.data(), 2*m_n_basins);
+  // copy values
+  result = result1;
+
+  for (int b = 1; b < 2 * m_n_basins; ++b) {
+    if (b % 2 == 0) {
+      m_log->message(2, "PICO, get basin neighbors of b=%d: b1=%d and b2=%d \n",
+                     b / 2, result[b], result[b + 1]);
+    }
+  }
+}
+
+/*!
+ *  Find all basins b, in which the ice shelf s has a calving front with potential ocean water intrusion.
+ *  Find the basin bmax, in which the ice shelf s has the most cells
+ */
+void PicoGeometry::identify_calving_front_connection(const IceModelVec2CellType &cell_type,
+                                                     const IceModelVec2Int &basin_mask,
+                                                     const IceModelVec2Int &shelf_mask,
+                                                     int n_shelves,
+                                                     std::vector<int> &most_shelf_cells_in_basin,
+                                                     std::vector<int> &cfs_in_basins_per_shelf) {
+
+  std::vector<int> n_shelf_cells_per_basin(n_shelves * m_n_basins,0);
+  // additional vectors to allreduce efficiently with IntelMPI
+  std::vector<int> n_shelf_cells_per_basinr(n_shelves * m_n_basins,0);
+  std::vector<int> cfs_in_basins_per_shelfr(n_shelves * m_n_basins,0);
+  std::vector<int> most_shelf_cells_in_basinr(n_shelves, 0);
+
+  IceModelVec::AccessList list{ &cell_type, &basin_mask, &shelf_mask };
+
+  {
+    for (Points p(*m_grid); p; p.next()) {
+      const int i = p.i(), j = p.j();
+      int s = shelf_mask.as_int(i, j);
+      int b = basin_mask.as_int(i, j);
+      int sb = s * m_n_basins + b;
+      n_shelf_cells_per_basin[sb]++;
+
+      if (cell_type.as_int(i, j) == MASK_FLOATING) {
+        auto M = cell_type.int_star(i, j);
+        if (M.n == MASK_ICE_FREE_OCEAN or
+            M.e == MASK_ICE_FREE_OCEAN or
+            M.s == MASK_ICE_FREE_OCEAN or
+            M.w == MASK_ICE_FREE_OCEAN) {
+          if (cfs_in_basins_per_shelf[sb] != b) {
+            cfs_in_basins_per_shelf[sb] = b;
+          }
+        }
+      }
+    }
+
+    GlobalSum(m_grid->com, cfs_in_basins_per_shelf.data(),
+              cfs_in_basins_per_shelfr.data(), n_shelves*m_n_basins);
+    GlobalSum(m_grid->com, n_shelf_cells_per_basin.data(),
+              n_shelf_cells_per_basinr.data(), n_shelves*m_n_basins);
+    // copy values
+    cfs_in_basins_per_shelf = cfs_in_basins_per_shelfr;
+    n_shelf_cells_per_basin = n_shelf_cells_per_basinr;
+
+    for (int s = 0; s < n_shelves; s++) {
+      int n_shelf_cells_per_basin_max = 0;
+      for (int b = 0; b < m_n_basins; b++) {
+        int sb = s * m_n_basins + b;
+        if (n_shelf_cells_per_basin[sb] > n_shelf_cells_per_basin_max) {
+          most_shelf_cells_in_basin[s] = b;
+          n_shelf_cells_per_basin_max  = n_shelf_cells_per_basin[sb];
+        }
+      }
+    }
+  }
+}
+
+
+/*!
+ * Find all ice shelves s that spread across non-neighboring basins with calving fronts in those basins and add an ice shelf mask number.
+ */
+void PicoGeometry::split_ice_shelves(const IceModelVec2CellType &cell_type,
+                                     const IceModelVec2Int &basin_mask,
+                                     const std::vector<int> &n_basin_neighbors,
+                                     const std::vector<int> &most_shelf_cells_in_basin,
+                                     const std::vector<int> &cfs_in_basins_per_shelf,
+                                     int n_shelves,
+                                     IceModelVec2Int &shelf_mask) {
+
+  m_tmp.copy_from(shelf_mask);
+
+  std::vector<int> n_shelf_cells_to_split(n_shelves * m_n_basins, 0);
+  std::vector<int> n_shelf_cells_to_splitr(n_shelves * m_n_basins, 0);
+
+  IceModelVec::AccessList list{ &cell_type, &basin_mask, &shelf_mask, &m_tmp };
+
+  for (Points p(*m_grid); p; p.next()) {
+    const int i = p.i(), j = p.j();
+    if (cell_type.as_int(i, j) == MASK_FLOATING) {
+      int b = basin_mask.as_int(i, j);
+      int s = shelf_mask.as_int(i, j);
+      int b0 = most_shelf_cells_in_basin[s];
+      if (b != b0 and
+          b != n_basin_neighbors[2 * b0] and
+          b != n_basin_neighbors[2 * b0 + 1] and
+          cfs_in_basins_per_shelf[s * m_n_basins + b] > 0) {
+        n_shelf_cells_to_split[s * m_n_basins + b]++;
+      }
+    }
+  }
+
+  GlobalSum(m_grid->com, n_shelf_cells_to_split.data(),
+            n_shelf_cells_to_splitr.data(), n_shelves * m_n_basins);
+  // copy values
+  n_shelf_cells_to_split = n_shelf_cells_to_splitr;
+
+  // no GlobalSum needed here, only local:
+  std::vector<int> add_shelf_instance(n_shelves * m_n_basins, 0);
+  int m_shelf_numbers_to_add = 0;
+  for (int s = 0; s < n_shelves; s++) {
+    int b0 = most_shelf_cells_in_basin[s];
+    for (int b = 0; b < m_n_basins; b++) {
+      if (n_shelf_cells_to_split[s * m_n_basins + b] > 0) {
+        m_shelf_numbers_to_add += 1;
+        add_shelf_instance[s * m_n_basins + b] = n_shelves + m_shelf_numbers_to_add;
+        m_log->message(3, "\nPICO, split ice shelf s=%d with bmax=%d "
+                       "and b=%d and n=%d and si=%d\n", s, b0, b,
+                       n_shelf_cells_to_split[s * m_n_basins + b],
+                       add_shelf_instance[s * m_n_basins + b]);
+      }
+    }
+  }
+
+  for (Points p(*m_grid); p; p.next()) {
+    const int i = p.i(), j = p.j();
+    if (cell_type.as_int(i, j) == MASK_FLOATING) {
+      int b = basin_mask.as_int(i, j);
+      int s = shelf_mask.as_int(i, j);
+      if (add_shelf_instance[s * m_n_basins + b] > 0) {
+        m_tmp(i, j) = add_shelf_instance[s * m_n_basins + b];
+      }
+    }
+  }
+
+  shelf_mask.copy_from(m_tmp);
+}
+
+/*!
+ * Compute distance to the grounding line.
+ */
 void PicoGeometry::compute_distances_gl(const IceModelVec2Int &ocean_mask,
                                         const IceModelVec2Int &ice_rises,
                                         bool exclude_ice_rises,
@@ -524,6 +752,9 @@ void PicoGeometry::compute_distances_gl(const IceModelVec2Int &ocean_mask,
   eikonal_equation(result);
 }
 
+/*!
+ * Compute distance to the calving front.
+ */
 void PicoGeometry::compute_distances_cf(const IceModelVec2Int &ocean_mask,
                                         const IceModelVec2Int &ice_rises,
                                         bool exclude_ice_rises,
@@ -580,6 +811,8 @@ void PicoGeometry::compute_distances_cf(const IceModelVec2Int &ocean_mask,
  * Note that this implementation updates ghosts *every* iteration. We could speed this
  * up by checking if a point at a boundary of the processor sub-domain was updated and
  * update ghosts in those cases only.
+ *
+ * FIXME: replace this with a better algorithm.
  */
 void eikonal_equation(IceModelVec2Int &mask) {
 
@@ -618,13 +851,17 @@ void eikonal_equation(IceModelVec2Int &mask) {
   }
 }
 
+/*!
+ * Compute the mask identifying ice shelf "boxes" using distances to the grounding line
+ * and the calving front.
+ */
 void PicoGeometry::compute_box_mask(const IceModelVec2Int &D_gl, const IceModelVec2Int &D_cf,
                                     const IceModelVec2Int &shelf_mask, int max_number_of_boxes,
                                     IceModelVec2Int &result) {
 
   IceModelVec::AccessList list{ &D_gl, &D_cf, &shelf_mask, &result };
 
-  int n_shelves = shelf_mask.range().max + 1;
+  int n_shelves = static_cast<int>(shelf_mask.range().max) + 1;
 
   std::vector<double> GL_distance_max(n_shelves, 0.0);
   std::vector<double> GL_distance_max1(n_shelves, 0.0);
@@ -634,7 +871,7 @@ void PicoGeometry::compute_box_mask(const IceModelVec2Int &D_gl, const IceModelV
   for (Points p(*m_grid); p; p.next()) {
     const int i = p.i(), j = p.j();
 
-    int shelf_id = shelf_mask(i, j);
+    int shelf_id = shelf_mask.as_int(i, j);
     assert(shelf_id >= 0);
     assert(shelf_id < n_shelves + 1);
 
@@ -664,8 +901,8 @@ void PicoGeometry::compute_box_mask(const IceModelVec2Int &D_gl, const IceModelV
   // compute the number of boxes in each shelf
 
   std::vector<int> n_boxes(n_shelves, 0);
-  int n_min   = 1;
-  double zeta = 0.5;
+  const int n_min   = 1;
+  const double zeta = 0.5;
 
   for (int k = 0; k < n_shelves; ++k) {
     n_boxes[k] = n_min + round(pow((GL_distance_max[k] / GL_distance_ref), zeta) * (max_number_of_boxes - n_min));
@@ -682,7 +919,7 @@ void PicoGeometry::compute_box_mask(const IceModelVec2Int &D_gl, const IceModelV
     int d_cf = D_cf.as_int(i, j);
 
     if (shelf_mask.as_int(i, j) > 0 and d_gl > 0.0 and d_cf > 0.0) {
-      int shelf_id = shelf_mask(i, j);
+      int shelf_id = shelf_mask.as_int(i, j);
       int n = n_boxes[shelf_id];
 
       // relative position on the shelf (ranges from 0 to 1), increasing towards the